+++ pkg/ddl/executor.go
@@ -79,7 +79,6 @@ import (
 	"github.com/pingcap/tidb/pkg/util/hack"
 	"github.com/pingcap/tidb/pkg/util/intest"
 	"github.com/pingcap/tidb/pkg/util/mathutil"
-	"github.com/pingcap/tidb/pkg/util/mock"
 	"github.com/pingcap/tidb/pkg/util/set"
 	"github.com/pingcap/tidb/pkg/util/sqlkiller"
 	"github.com/pingcap/tidb/pkg/util/stringutil"
@@ -884,117 +883,6 @@ func checkTooLongForeignKey(fk model.CIStr) error {
 		return dbterror.ErrTooLongIdent.GenWithStackByArgs(fk)
 	}
 	return nil
-}
-
-func setColumnFlagWithConstraint(colMap map[string]*table.Column, v *ast.Constraint) {
-	switch v.Tp {
-	case ast.ConstraintPrimaryKey:
-		for _, key := range v.Keys {
-			if key.Expr != nil {
-				continue
-			}
-			c, ok := colMap[key.Column.Name.L]
-			if !ok {
-				continue
-			}
-			c.AddFlag(mysql.PriKeyFlag)
-			// Primary key can not be NULL.
-			c.AddFlag(mysql.NotNullFlag)
-			setNoDefaultValueFlag(c, c.DefaultValue != nil)
-		}
-	case ast.ConstraintUniq, ast.ConstraintUniqIndex, ast.ConstraintUniqKey:
-		for i, key := range v.Keys {
-			if key.Expr != nil {
-				continue
-			}
-			c, ok := colMap[key.Column.Name.L]
-			if !ok {
-				continue
-			}
-			if i == 0 {
-				// Only the first column can be set
-				// if unique index has multi columns,
-				// the flag should be MultipleKeyFlag.
-				// See https://dev.mysql.com/doc/refman/5.7/en/show-columns.html
-				if len(v.Keys) > 1 {
-					c.AddFlag(mysql.MultipleKeyFlag)
-				} else {
-					c.AddFlag(mysql.UniqueKeyFlag)
-				}
-			}
-		}
-	case ast.ConstraintKey, ast.ConstraintIndex:
-		for i, key := range v.Keys {
-			if key.Expr != nil {
-				continue
-			}
-			c, ok := colMap[key.Column.Name.L]
-			if !ok {
-				continue
-			}
-			if i == 0 {
-				// Only the first column can be set.
-				c.AddFlag(mysql.MultipleKeyFlag)
-			}
-		}
-	}
-}
-
-func buildColumnsAndConstraints(
-	ctx sessionctx.Context,
-	colDefs []*ast.ColumnDef,
-	constraints []*ast.Constraint,
-	tblCharset string,
-	tblCollate string,
-) ([]*table.Column, []*ast.Constraint, error) {
-	// outPriKeyConstraint is the primary key constraint out of column definition. such as: create table t1 (id int , age int, primary key(id));
-	var outPriKeyConstraint *ast.Constraint
-	for _, v := range constraints {
-		if v.Tp == ast.ConstraintPrimaryKey {
-			outPriKeyConstraint = v
-			break
-		}
-	}
-	cols := make([]*table.Column, 0, len(colDefs))
-	colMap := make(map[string]*table.Column, len(colDefs))
-
-	for i, colDef := range colDefs {
-		if field_types.TiDBStrictIntegerDisplayWidth {
-			switch colDef.Tp.GetType() {
-			case mysql.TypeTiny:
-				// No warning for BOOL-like tinyint(1)
-				if colDef.Tp.GetFlen() != types.UnspecifiedLength && colDef.Tp.GetFlen() != 1 {
-					ctx.GetSessionVars().StmtCtx.AppendWarning(
-						dbterror.ErrWarnDeprecatedIntegerDisplayWidth.FastGenByArgs(),
-					)
-				}
-			case mysql.TypeShort, mysql.TypeInt24, mysql.TypeLong, mysql.TypeLonglong:
-				if colDef.Tp.GetFlen() != types.UnspecifiedLength {
-					ctx.GetSessionVars().StmtCtx.AppendWarning(
-						dbterror.ErrWarnDeprecatedIntegerDisplayWidth.FastGenByArgs(),
-					)
-				}
-			}
-		}
-		col, cts, err := buildColumnAndConstraint(ctx, i, colDef, outPriKeyConstraint, tblCharset, tblCollate)
-		if err != nil {
-			return nil, nil, errors.Trace(err)
-		}
-		col.State = model.StatePublic
-		if mysql.HasZerofillFlag(col.GetFlag()) {
-			ctx.GetSessionVars().StmtCtx.AppendWarning(
-				dbterror.ErrWarnDeprecatedZerofill.FastGenByArgs(),
-			)
-		}
-		constraints = append(constraints, cts...)
-		cols = append(cols, col)
-		colMap[colDef.Name.Name.L] = col
-	}
-	// Traverse table Constraints and set col.flag.
-	for _, v := range constraints {
-		setColumnFlagWithConstraint(colMap, v)
-	}
-	return cols, constraints, nil
 }
 
 func getDefaultCollationForUTF8MB4(sessVars *variable.SessionVars, cs string) (string, error) {
@@ -2192,344 +2080,6 @@ func checkInvisibleIndexOnPK(tblInfo *model.TableInfo) error {
 	return nil
 }
 
-func setTableAutoRandomBits(ctx sessionctx.Context, tbInfo *model.TableInfo, colDefs []*ast.ColumnDef) error {
-	for _, col := range colDefs {
-		if containsColumnOption(col, ast.ColumnOptionAutoRandom) {
-			if col.Tp.GetType() != mysql.TypeLonglong {
-				return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(
-					fmt.Sprintf(autoid.AutoRandomOnNonBigIntColumn, types.TypeStr(col.Tp.GetType())))
-			}
-			switch {
-			case tbInfo.PKIsHandle:
-				if tbInfo.GetPkName().L != col.Name.Name.L {
-					errMsg := fmt.Sprintf(autoid.AutoRandomMustFirstColumnInPK, col.Name.Name.O)
-					return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(errMsg)
-				}
-			case tbInfo.IsCommonHandle:
-				pk := tables.FindPrimaryIndex(tbInfo)
-				if pk == nil {
-					return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(autoid.AutoRandomNoClusteredPKErrMsg)
-				}
-				if col.Name.Name.L != pk.Columns[0].Name.L {
-					errMsg := fmt.Sprintf(autoid.AutoRandomMustFirstColumnInPK, col.Name.Name.O)
-					return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(errMsg)
-				}
-			default:
-				return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(autoid.AutoRandomNoClusteredPKErrMsg)
-			}
-
-			if containsColumnOption(col, ast.ColumnOptionAutoIncrement) {
-				return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(autoid.AutoRandomIncompatibleWithAutoIncErrMsg)
-			}
-			if containsColumnOption(col, ast.ColumnOptionDefaultValue) {
-				return dbterror.ErrInvalidAutoRandom.GenWithStackByArgs(autoid.AutoRandomIncompatibleWithDefaultValueErrMsg)
-			}
-
-			shardBits, rangeBits, err := extractAutoRandomBitsFromColDef(col)
-			if err != nil {
-				return errors.Trace(err)
-			}
-			tbInfo.AutoRandomBits = shardBits
-			tbInfo.AutoRandomRangeBits = rangeBits
-
-			shardFmt := autoid.NewShardIDFormat(col.Tp, shardBits, rangeBits)
-			if shardFmt.IncrementalBits < autoid.AutoRandomIncBitsMin {
-				return dbterror.ErrInvalidAutoRandom.FastGenByArgs(autoid.AutoRandomIncrementalBitsTooSmall)
-			}
-			msg := fmt.Sprintf(autoid.AutoRandomAvailableAllocTimesNote, shardFmt.IncrementalBitsCapacity())
-			ctx.GetSessionVars().StmtCtx.AppendNote(errors.NewNoStackError(msg))
-		}
-	}
-	return nil
-}
-
-func extractAutoRandomBitsFromColDef(colDef *ast.ColumnDef) (shardBits, rangeBits uint64, err error) {
-	for _, op := range colDef.Options {
-		if op.Tp == ast.ColumnOptionAutoRandom {
-			shardBits, err = autoid.AutoRandomShardBitsNormalize(op.AutoRandOpt.ShardBits, colDef.Name.Name.O)
-			if err != nil {
-				return 0, 0, err
-			}
-			rangeBits, err = autoid.AutoRandomRangeBitsNormalize(op.AutoRandOpt.RangeBits)
-			if err != nil {
-				return 0, 0, err
-			}
-			return shardBits, rangeBits, nil
-		}
-	}
-	return 0, 0, nil
-}
-
-// BuildTableInfo creates a TableInfo.
-func BuildTableInfo(
-	ctx sessionctx.Context,
-	tableName model.CIStr,
-	cols []*table.Column,
-	constraints []*ast.Constraint,
-	charset string,
-	collate string,
-) (tbInfo *model.TableInfo, err error) {
-	tbInfo = &model.TableInfo{
-		Name:    tableName,
-		Version: model.CurrLatestTableInfoVersion,
-		Charset: charset,
-		Collate: collate,
-	}
-	tblColumns := make([]*table.Column, 0, len(cols))
-	existedColsMap := make(map[string]struct{}, len(cols))
-	for _, v := range cols {
-		v.ID = AllocateColumnID(tbInfo)
-		tbInfo.Columns = append(tbInfo.Columns, v.ToInfo())
-		tblColumns = append(tblColumns, table.ToColumn(v.ToInfo()))
-		existedColsMap[v.Name.L] = struct{}{}
-	}
-	foreignKeyID := tbInfo.MaxForeignKeyID
-	for _, constr := range constraints {
-		// Build hidden columns if necessary.
-		hiddenCols, err := buildHiddenColumnInfoWithCheck(ctx, constr.Keys, model.NewCIStr(constr.Name), tbInfo, tblColumns)
-		if err != nil {
-			return nil, err
-		}
-		for _, hiddenCol := range hiddenCols {
-			hiddenCol.State = model.StatePublic
-			hiddenCol.ID = AllocateColumnID(tbInfo)
-			hiddenCol.Offset = len(tbInfo.Columns)
-			tbInfo.Columns = append(tbInfo.Columns, hiddenCol)
-			tblColumns = append(tblColumns, table.ToColumn(hiddenCol))
-		}
-		// Check clustered on non-primary key.
-		if constr.Option != nil && constr.Option.PrimaryKeyTp != model.PrimaryKeyTypeDefault &&
-			constr.Tp != ast.ConstraintPrimaryKey {
-			return nil, dbterror.ErrUnsupportedClusteredSecondaryKey
-		}
-		if constr.Tp == ast.ConstraintForeignKey {
-			var fkName model.CIStr
-			foreignKeyID++
-			if constr.Name != "" {
-				fkName = model.NewCIStr(constr.Name)
-			} else {
-				fkName = model.NewCIStr(fmt.Sprintf("fk_%d", foreignKeyID))
-			}
-			if model.FindFKInfoByName(tbInfo.ForeignKeys, fkName.L) != nil {
-				return nil, infoschema.ErrCannotAddForeign
-			}
-			fk, err := buildFKInfo(fkName, constr.Keys, constr.Refer, cols)
-			if err != nil {
-				return nil, err
-			}
-			fk.State = model.StatePublic
-
-			tbInfo.ForeignKeys = append(tbInfo.ForeignKeys, fk)
-			continue
-		}
-		if constr.Tp == ast.ConstraintPrimaryKey {
-			lastCol, err := CheckPKOnGeneratedColumn(tbInfo, constr.Keys)
-			if err != nil {
-				return nil, err
-			}
-			isSingleIntPK := isSingleIntPK(constr, lastCol)
-			if ShouldBuildClusteredIndex(ctx, constr.Option, isSingleIntPK) {
-				if isSingleIntPK {
-					tbInfo.PKIsHandle = true
-				} else {
-					tbInfo.IsCommonHandle = true
-					tbInfo.CommonHandleVersion = 1
-				}
-			}
-			if tbInfo.HasClusteredIndex() {
-				// Primary key cannot be invisible.
-				if constr.Option != nil && constr.Option.Visibility == ast.IndexVisibilityInvisible {
-					return nil, dbterror.ErrPKIndexCantBeInvisible
-				}
-			}
-			if tbInfo.PKIsHandle {
-				continue
-			}
-		}
-
-		if constr.Tp == ast.ConstraintFulltext {
-			ctx.GetSessionVars().StmtCtx.AppendWarning(dbterror.ErrTableCantHandleFt.FastGenByArgs())
-			continue
-		}
-
-		var (
-			indexName       = constr.Name
-			primary, unique bool
-		)
-
-		// Check if the index is primary or unique.
-		switch constr.Tp {
-		case ast.ConstraintPrimaryKey:
-			primary = true
-			unique = true
-			indexName = mysql.PrimaryKeyName
-		case ast.ConstraintUniq, ast.ConstraintUniqKey, ast.ConstraintUniqIndex:
-			unique = true
-		}
-
-		// check constraint
-		if constr.Tp == ast.ConstraintCheck {
-			if !variable.EnableCheckConstraint.Load() {
-				ctx.GetSessionVars().StmtCtx.AppendWarning(errCheckConstraintIsOff)
-				continue
-			}
-			// Since column check constraint dependency has been done in columnDefToCol.
-			// Here do the table check constraint dependency check, table constraint
-			// can only refer the columns in defined columns of the table.
-			// Refer: https://dev.mysql.com/doc/refman/8.0/en/create-table-check-constraints.html
-			if ok, err := table.IsSupportedExpr(constr); !ok {
-				return nil, err
-			}
-			var dependedCols []model.CIStr
-			dependedColsMap := findDependentColsInExpr(constr.Expr)
-			if !constr.InColumn {
-				dependedCols = make([]model.CIStr, 0, len(dependedColsMap))
-				for k := range dependedColsMap {
-					if _, ok := existedColsMap[k]; !ok {
-						// The table constraint depended on a non-existed column.
-						return nil, dbterror.ErrTableCheckConstraintReferUnknown.GenWithStackByArgs(constr.Name, k)
-					}
-					dependedCols = append(dependedCols, model.NewCIStr(k))
-				}
-			} else {
-				// Check the column-type constraint dependency.
-				if len(dependedColsMap) > 1 {
-					return nil, dbterror.ErrColumnCheckConstraintReferOther.GenWithStackByArgs(constr.Name)
-				} else if len(dependedColsMap) == 0 {
-					// If dependedCols is empty, the expression must be true/false.
-					valExpr, ok := constr.Expr.(*driver.ValueExpr)
-					if !ok || !mysql.HasIsBooleanFlag(valExpr.GetType().GetFlag()) {
-						return nil, errors.Trace(errors.New("unsupported expression in check constraint"))
-					}
-				} else {
-					if _, ok := dependedColsMap[constr.InColumnName]; !ok {
-						return nil, dbterror.ErrColumnCheckConstraintReferOther.GenWithStackByArgs(constr.Name)
-					}
-					dependedCols = []model.CIStr{model.NewCIStr(constr.InColumnName)}
-				}
-			}
-			// check auto-increment column
-			if table.ContainsAutoIncrementCol(dependedCols, tbInfo) {
-				return nil, dbterror.ErrCheckConstraintRefersAutoIncrementColumn.GenWithStackByArgs(constr.Name)
-			}
-			// check foreign key
-			if err := table.HasForeignKeyRefAction(tbInfo.ForeignKeys, constraints, constr, dependedCols); err != nil {
-				return nil, err
-			}
-			// build constraint meta info.
-			constraintInfo, err := buildConstraintInfo(tbInfo, dependedCols, constr, model.StatePublic)
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			// check if the expression is bool type
-			if err := table.IfCheckConstraintExprBoolType(ctx.GetExprCtx().GetEvalCtx(), constraintInfo, tbInfo); err != nil {
-				return nil, err
-			}
-			constraintInfo.ID = allocateConstraintID(tbInfo)
-			tbInfo.Constraints = append(tbInfo.Constraints, constraintInfo)
-			continue
-		}
-
-		// build index info.
-		idxInfo, err := BuildIndexInfo(
-			ctx,
-			tbInfo.Columns,
-			model.NewCIStr(indexName),
-			primary,
-			unique,
-			false,
-			constr.Keys,
-			constr.Option,
-			model.StatePublic,
-		)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-
-		if len(hiddenCols) > 0 {
-			AddIndexColumnFlag(tbInfo, idxInfo)
-		}
-		sessionVars := ctx.GetSessionVars()
-		_, err = validateCommentLength(sessionVars.StmtCtx.ErrCtx(), sessionVars.SQLMode, idxInfo.Name.String(), &idxInfo.Comment, dbterror.ErrTooLongIndexComment)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		idxInfo.ID = AllocateIndexID(tbInfo)
-		tbInfo.Indices = append(tbInfo.Indices, idxInfo)
-	}
-
-	err = addIndexForForeignKey(ctx, tbInfo)
-	return tbInfo, err
-}
-
-// addIndexForForeignKey uses to auto create an index for the foreign key if the table doesn't have any index cover the
-// foreign key columns.
-func addIndexForForeignKey(ctx sessionctx.Context, tbInfo *model.TableInfo) error {
-	if len(tbInfo.ForeignKeys) == 0 {
-		return nil
-	}
-	var handleCol *model.ColumnInfo
-	if tbInfo.PKIsHandle {
-		handleCol = tbInfo.GetPkColInfo()
-	}
-	for _, fk := range tbInfo.ForeignKeys {
-		if fk.Version < model.FKVersion1 {
-			continue
-		}
-		if handleCol != nil && len(fk.Cols) == 1 && handleCol.Name.L == fk.Cols[0].L {
-			continue
-		}
-		if model.FindIndexByColumns(tbInfo, tbInfo.Indices, fk.Cols...) != nil {
-			continue
-		}
-		idxName := fk.Name
-		if tbInfo.FindIndexByName(idxName.L) != nil {
-			return dbterror.ErrDupKeyName.GenWithStack("duplicate key name %s", fk.Name.O)
-		}
-		keys := make([]*ast.IndexPartSpecification, 0, len(fk.Cols))
-		for _, col := range fk.Cols {
-			keys = append(keys, &ast.IndexPartSpecification{
-				Column: &ast.ColumnName{Name: col},
-				Length: types.UnspecifiedLength,
-			})
-		}
-		idxInfo, err := BuildIndexInfo(ctx, tbInfo.Columns, idxName, false, false, false, keys, nil, model.StatePublic)
-		if err != nil {
-			return errors.Trace(err)
-		}
-		idxInfo.ID = AllocateIndexID(tbInfo)
-		tbInfo.Indices = append(tbInfo.Indices, idxInfo)
-	}
-	return nil
-}
-
-func isSingleIntPK(constr *ast.Constraint, lastCol *model.ColumnInfo) bool {
-	if len(constr.Keys) != 1 {
-		return false
-	}
-	switch lastCol.GetType() {
-	case mysql.TypeLong, mysql.TypeLonglong,
-		mysql.TypeTiny, mysql.TypeShort, mysql.TypeInt24:
-		return true
-	}
-	return false
-}
-
-// ShouldBuildClusteredIndex is used to determine whether the CREATE TABLE statement should build a clustered index table.
-func ShouldBuildClusteredIndex(ctx sessionctx.Context, opt *ast.IndexOption, isSingleIntPK bool) bool {
-	if opt == nil || opt.PrimaryKeyTp == model.PrimaryKeyTypeDefault {
-		switch ctx.GetSessionVars().EnableClusteredIndex {
-		case variable.ClusteredIndexDefModeOn:
-			return true
-		case variable.ClusteredIndexDefModeIntOnly:
-			return !config.GetGlobalConfig().AlterPrimaryKey && isSingleIntPK
-		default:
-			return false
-		}
-	}
-	return opt.PrimaryKeyTp == model.PrimaryKeyTypeClustered
-}
-
 // checkTableInfoValidExtra is like checkTableInfoValid, but also assumes the
 // table info comes from untrusted source and performs further checks such as
 // name length and column count.
@@ -2641,171 +2191,6 @@ func checkTableInfoValid(tblInfo *model.TableInfo) error {
 	return checkInvisibleIndexOnPK(tblInfo)
 }
 
-// BuildTableInfoWithLike builds a new table info according to CREATE TABLE ... LIKE statement.
-func BuildTableInfoWithLike(ctx sessionctx.Context, ident ast.Ident, referTblInfo *model.TableInfo, s *ast.CreateTableStmt) (*model.TableInfo, error) {
-	// Check the referred table is a real table object.
-	if referTblInfo.IsSequence() || referTblInfo.IsView() {
-		return nil, dbterror.ErrWrongObject.GenWithStackByArgs(ident.Schema, referTblInfo.Name, "BASE TABLE")
-	}
-	tblInfo := *referTblInfo
-	if err := setTemporaryType(ctx, &tblInfo, s); err != nil {
-		return nil, errors.Trace(err)
-	}
-	// Check non-public column and adjust column offset.
-	newColumns := referTblInfo.Cols()
-	newIndices := make([]*model.IndexInfo, 0, len(tblInfo.Indices))
-	for _, idx := range tblInfo.Indices {
-		if idx.State == model.StatePublic {
-			newIndices = append(newIndices, idx)
-		}
-	}
-	tblInfo.Columns = newColumns
-	tblInfo.Indices = newIndices
-	tblInfo.Name = ident.Name
-	tblInfo.AutoIncID = 0
-	tblInfo.ForeignKeys = nil
-	// Ignore TiFlash replicas for temporary tables.
-	if s.TemporaryKeyword != ast.TemporaryNone {
-		tblInfo.TiFlashReplica = nil
-	} else if tblInfo.TiFlashReplica != nil {
-		replica := *tblInfo.TiFlashReplica
-		// Keep the tiflash replica setting, remove the replica available status.
-		replica.AvailablePartitionIDs = nil
-		replica.Available = false
-		tblInfo.TiFlashReplica = &replica
-	}
-	if referTblInfo.Partition != nil {
-		pi := *referTblInfo.Partition
-		pi.Definitions = make([]model.PartitionDefinition, len(referTblInfo.Partition.Definitions))
-		copy(pi.Definitions, referTblInfo.Partition.Definitions)
-		tblInfo.Partition = &pi
-	}
-
-	if referTblInfo.TTLInfo != nil {
-		tblInfo.TTLInfo = referTblInfo.TTLInfo.Clone()
-	}
-	renameCheckConstraint(&tblInfo)
-	return &tblInfo, nil
-}
-
-func renameCheckConstraint(tblInfo *model.TableInfo) {
-	for _, cons := range tblInfo.Constraints {
-		cons.Name = model.NewCIStr("")
-		cons.Table = tblInfo.Name
-	}
-	setNameForConstraintInfo(tblInfo.Name.L, map[string]bool{}, tblInfo.Constraints)
-}
-
-// BuildTableInfoFromAST builds model.TableInfo from a SQL statement.
-// Note: TableID and PartitionID are left as uninitialized value.
-func BuildTableInfoFromAST(s *ast.CreateTableStmt) (*model.TableInfo, error) {
-	return buildTableInfoWithCheck(mock.NewContext(), s, mysql.DefaultCharset, "", nil)
-}
-
-// buildTableInfoWithCheck builds model.TableInfo from a SQL statement.
-// Note: TableID and PartitionIDs are left as uninitialized value.
-func buildTableInfoWithCheck(ctx sessionctx.Context, s *ast.CreateTableStmt, dbCharset, dbCollate string, placementPolicyRef *model.PolicyRefInfo) (*model.TableInfo, error) {
-	tbInfo, err := BuildTableInfoWithStmt(ctx, s, dbCharset, dbCollate, placementPolicyRef)
-	if err != nil {
-		return nil, err
-	}
-	// Fix issue 17952 which will cause partition range expr can't be parsed as Int.
-	// checkTableInfoValidWithStmt will do the constant fold the partition expression first,
-	// then checkTableInfoValidExtra will pass the tableInfo check successfully.
-	if err = checkTableInfoValidWithStmt(ctx, tbInfo, s); err != nil {
-		return nil, err
-	}
-	if err = checkTableInfoValidExtra(tbInfo); err != nil {
-		return nil, err
-	}
-	return tbInfo, nil
-}
-
-// BuildSessionTemporaryTableInfo builds model.TableInfo from a SQL statement.
-func BuildSessionTemporaryTableInfo(ctx sessionctx.Context, is infoschema.InfoSchema, s *ast.CreateTableStmt, dbCharset, dbCollate string, placementPolicyRef *model.PolicyRefInfo) (*model.TableInfo, error) {
-	ident := ast.Ident{Schema: s.Table.Schema, Name: s.Table.Name}
-	//build tableInfo
-	var tbInfo *model.TableInfo
-	var referTbl table.Table
-	var err error
-	if s.ReferTable != nil {
-		referIdent := ast.Ident{Schema: s.ReferTable.Schema, Name: s.ReferTable.Name}
-		_, ok := is.SchemaByName(referIdent.Schema)
-		if !ok {
-			return nil, infoschema.ErrTableNotExists.GenWithStackByArgs(referIdent.Schema, referIdent.Name)
-		}
-		referTbl, err = is.TableByName(context.Background(), referIdent.Schema, referIdent.Name)
-		if err != nil {
-			return nil, infoschema.ErrTableNotExists.GenWithStackByArgs(referIdent.Schema, referIdent.Name)
-		}
-		tbInfo, err = BuildTableInfoWithLike(ctx, ident, referTbl.Meta(), s)
-	} else {
-		tbInfo, err = buildTableInfoWithCheck(ctx, s, dbCharset, dbCollate, placementPolicyRef)
-	}
-	return tbInfo, err
-}
-
-// BuildTableInfoWithStmt builds model.TableInfo from a SQL statement without validity check
-func BuildTableInfoWithStmt(ctx sessionctx.Context, s *ast.CreateTableStmt, dbCharset, dbCollate string, placementPolicyRef *model.PolicyRefInfo) (*model.TableInfo, error) {
-	colDefs := s.Cols
-	tableCharset, tableCollate, err := GetCharsetAndCollateInTableOption(ctx.GetSessionVars(), 0, s.Options)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	tableCharset, tableCollate, err = ResolveCharsetCollation(ctx.GetSessionVars(),
-		ast.CharsetOpt{Chs: tableCharset, Col: tableCollate},
-		ast.CharsetOpt{Chs: dbCharset, Col: dbCollate},
-	)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	// The column charset haven't been resolved here.
-	cols, newConstraints, err := buildColumnsAndConstraints(ctx, colDefs, s.Constraints, tableCharset, tableCollate)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	err = checkConstraintNames(s.Table.Name, newConstraints)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	var tbInfo *model.TableInfo
-	tbInfo, err = BuildTableInfo(ctx, s.Table.Name, cols, newConstraints, tableCharset, tableCollate)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if err = setTemporaryType(ctx, tbInfo, s); err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err = setTableAutoRandomBits(ctx, tbInfo, colDefs); err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if err = handleTableOptions(s.Options, tbInfo); err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	sessionVars := ctx.GetSessionVars()
-	if _, err = validateCommentLength(sessionVars.StmtCtx.ErrCtx(), sessionVars.SQLMode, tbInfo.Name.L, &tbInfo.Comment, dbterror.ErrTooLongTableComment); err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if tbInfo.TempTableType == model.TempTableNone && tbInfo.PlacementPolicyRef == nil && placementPolicyRef != nil {
-		// Set the defaults from Schema. Note: they are mutual exclusive!
-		tbInfo.PlacementPolicyRef = placementPolicyRef
-	}
-
-	// After handleTableOptions, so the partitions can get defaults from Table level
-	err = buildTablePartitionInfo(ctx, s.Partition, tbInfo)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	return tbInfo, nil
-}
-
 func (e *executor) assignPartitionIDs(defs []model.PartitionDefinition) error {
 	genIDs, err := e.genGlobalIDs(len(defs))
 	if err != nil {
@@ -2870,22 +2255,6 @@ func (e *executor) CreateTable(ctx sessionctx.Context, s *ast.CreateTableStmt) (
 	}
 
 	return e.CreateTableWithInfo(ctx, schema.Name, tbInfo, involvingRef, WithOnExist(onExist))
-}
-
-func setTemporaryType(_ sessionctx.Context, tbInfo *model.TableInfo, s *ast.CreateTableStmt) error {
-	switch s.TemporaryKeyword {
-	case ast.TemporaryGlobal:
-		tbInfo.TempTableType = model.TempTableGlobal
-		// "create global temporary table ... on commit preserve rows"
-		if !s.OnCommitDelete {
-			return errors.Trace(dbterror.ErrUnsupportedOnCommitPreserve)
-		}
-	case ast.TemporaryLocal:
-		tbInfo.TempTableType = model.TempTableLocal
-	default:
-		tbInfo.TempTableType = model.TempTableNone
-	}
-	return nil
 }
 
 // createTableWithInfoJob returns the table creation job.
@@ -3817,77 +3186,6 @@ func parseBackgroundJobTypes(t string) ([]string, error) {
 		}
 	}
 	return res, nil
-}
-
-// handleTableOptions updates tableInfo according to table options.
-func handleTableOptions(options []*ast.TableOption, tbInfo *model.TableInfo) error {
-	var ttlOptionsHandled bool
-
-	for _, op := range options {
-		switch op.Tp {
-		case ast.TableOptionAutoIncrement:
-			tbInfo.AutoIncID = int64(op.UintValue)
-		case ast.TableOptionAutoIdCache:
-			if op.UintValue > uint64(math.MaxInt64) {
-				// TODO: Refine this error.
-				return errors.New("table option auto_id_cache overflows int64")
-			}
-			tbInfo.AutoIdCache = int64(op.UintValue)
-		case ast.TableOptionAutoRandomBase:
-			tbInfo.AutoRandID = int64(op.UintValue)
-		case ast.TableOptionComment:
-			tbInfo.Comment = op.StrValue
-		case ast.TableOptionCompression:
-			tbInfo.Compression = op.StrValue
-		case ast.TableOptionShardRowID:
-			if op.UintValue > 0 && tbInfo.HasClusteredIndex() {
-				return dbterror.ErrUnsupportedShardRowIDBits
-			}
-			tbInfo.ShardRowIDBits = op.UintValue
-			if tbInfo.ShardRowIDBits > shardRowIDBitsMax {
-				tbInfo.ShardRowIDBits = shardRowIDBitsMax
-			}
-			tbInfo.MaxShardRowIDBits = tbInfo.ShardRowIDBits
-		case ast.TableOptionPreSplitRegion:
-			if tbInfo.TempTableType != model.TempTableNone {
-				return errors.Trace(dbterror.ErrOptOnTemporaryTable.GenWithStackByArgs("pre split regions"))
-			}
-			tbInfo.PreSplitRegions = op.UintValue
-		case ast.TableOptionCharset, ast.TableOptionCollate:
-			// We don't handle charset and collate here since they're handled in `GetCharsetAndCollateInTableOption`.
-		case ast.TableOptionPlacementPolicy:
-			tbInfo.PlacementPolicyRef = &model.PolicyRefInfo{
-				Name: model.NewCIStr(op.StrValue),
-			}
-		case ast.TableOptionTTL, ast.TableOptionTTLEnable, ast.TableOptionTTLJobInterval:
-			if ttlOptionsHandled {
-				continue
-			}
-
-			ttlInfo, ttlEnable, ttlJobInterval, err := getTTLInfoInOptions(options)
-			if err != nil {
-				return err
-			}
-			// It's impossible that `ttlInfo` and `ttlEnable` are all nil, because we have met this option.
-			// After exclude the situation `ttlInfo == nil && ttlEnable != nil`, we could say `ttlInfo != nil`
-			if ttlInfo == nil {
-				if ttlEnable != nil {
-					return errors.Trace(dbterror.ErrSetTTLOptionForNonTTLTable.FastGenByArgs("TTL_ENABLE"))
-				}
-				if ttlJobInterval != nil {
-					return errors.Trace(dbterror.ErrSetTTLOptionForNonTTLTable.FastGenByArgs("TTL_JOB_INTERVAL"))
-				}
-			}
-
-			tbInfo.TTLInfo = ttlInfo
-			ttlOptionsHandled = true
-		}
-	}
-	shardingBits := shardingBits(tbInfo)
-	if tbInfo.PreSplitRegions > shardingBits {
-		tbInfo.PreSplitRegions = shardingBits
-	}
-	return nil
 }
 
 func shardingBits(tblInfo *model.TableInfo) uint64 {
